requirements.txt
Django==5.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.14.0
psycopg2-binary==2.9.5
python-dotenv==1.0.0
Pillow==9.4.0
celery==5.2.7
redis==4.5.4
gunicorn==20.1.0
django-filter==23.1 